Move 'jrolfs' layout code into... layout
