--- keyboards/lets_split/keymaps/jrolfs/config.h
@@ -1,14 +1,0 @@
-#ifndef CONFIG_USER_H
-#define CONFIG_USER_H
-
-#include "../../config.h"
-
-#define USE_SERIAL
-
-#undef RGBLED_NUM
-#define RGBLED_NUM 8
-#define RGBLIGHT_ANIMATIONS
-
-#define EE_HANDS
-
-#endif
