--- keyboards/lets_split/keymaps/jrolfs/keymap.c
@@ -1,4 +1,0 @@
-#include "lets_split.h"
-
-#include "jrolfs.h"
-#include "jrolfs_layers.h"
