--- keyboards/lets_split/keymaps/jrolfs/rules.mk
@@ -1,5 +1,0 @@
-RGBLIGHT_ENABLE = yes
-
-ifndef QUANTUM_DIR
-	include ../../../../Makefile
-endif
