--- keyboards/niu_mini/keymaps/jrolfs/keymap.c
@@ -1,4 +1,0 @@
-#include "niu_mini.h"
-
-#include "jrolfs.h"
-#include "jrolfs_layers.h"
