--- keyboards/planck/keymaps/jrolfs/config.h
@@ -1,8 +1,0 @@
-#ifndef CONFIG_USER_H
-#define CONFIG_USER_H
-
-#include "../../config.h"
-
-#define USB_MAX_POWER_CONSUMPTION 50
-
-#endif
