--- keyboards/planck/keymaps/jrolfs/keymap.c
@@ -1,4 +1,0 @@
-#include "planck.h"
-
-#include "jrolfs.h"
-#include "jrolfs_layers.h"
