+++ layouts/community/ortho_4x12/jrolfs/config.h
@@ -1,0 +1,6 @@
+#ifndef CONFIG_USER_H
+#define CONFIG_USER_H
+
+#include QMK_KEYBOARD_CONFIG_H
+
+#endif
